requirements.txt
py-cord[voice]>=1.7.3

nest_asyncio==1.5.1
fuzzywuzzy==0.18.0

ormar==0.10.15
aiomysql>=0.0.21
mysqlclient>=1.7.4
alembic>=1.7.4

Hypercorn==0.11.2
Quart==0.17
Quart_Discord==2.1.4

moviepy==1.0.3
numpy==1.21.1
array2gif==1.0.4

pytest>=6.2.5
pytest-env>=0.6.2
